Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
passlib
bcrypt
python-jose[cryptography]
python-multipart